feat: move more portable types to commons (batch 20 part 3)
Additional moves:
- SupportedContent (video content filter)
- TorServiceStatus (tor status sealed class)
- StringFeedState (settings feed state)
- ServerName, ServerType, DEFAULT_MEDIA_SERVERS (media server config)

Part of the KMP iOS migration tracked in #2238.
