+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/settings/StringFeedState.kt
@@ -20,19 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.settings
 
-import kotlinx.collections.immutable.ImmutableList
-import kotlinx.coroutines.flow.MutableStateFlow
-
-sealed class StringFeedState {
-    object Loading : StringFeedState()
-
-    class Loaded(
-        val feed: MutableStateFlow<ImmutableList<String>>,
-    ) : StringFeedState()
-
-    object Empty : StringFeedState()
-
-    class FeedError(
-        val errorMessage: String,
-    ) : StringFeedState()
-}
+// Re-export from commons for backwards compatibility
+typealias StringFeedState = com.vitorpamplona.amethyst.commons.ui.screens.settings.StringFeedState
